Make a class for creating context
