+++ packages/moleculer-graphql/src/classes/index.ts
@@ -1,4 +1,10 @@
 export { default as GatewayStitcher } from './GatewayStitcher';
+export {
+	default as GraphQLContextCreator,
+	type GraphQLContext,
+	type GraphQLContextBase,
+	type GraphQLContextFactory,
+} from './GraphQLContextCreator';
 export { default as GraphQLExecutor } from './GraphQLExecutor';
 export { default as RequestHandler, type Request } from './RequestHandler';
 export { default as SchemaBuilder, type SchemaDirectiveTransformer } from './SchemaBuilder';
