+++ packages/moleculer-graphql/src/functions/index.ts
@@ -1,7 +1,1 @@
-export {
-	default as createGraphQLContext,
-	type GraphQLContextFactory,
-	type GraphQLContextBase,
-	type GraphQLContext,
-} from './createGraphQLContext';
 export { default as createValidate } from './createValidate';
