+++ packages/moleculer-graphql/src/mixins/serviceMixin.ts
@@ -2,9 +2,8 @@ import type { SubschemaConfig } from '@graphql-tools/delegate';
 import type { IResolvers } from '@graphql-tools/utils';
 import { defaultsDeep } from 'lodash';
 import type { Context, Service, ServiceSchema, ServiceSettingSchema } from 'moleculer';
-import type { SchemaDirectiveTransformer } from '../classes';
 import { GraphQLExecutor, SchemaBuilder } from '../classes';
-import type { GraphQLContext, GraphQLContextFactory } from '../functions';
+import type { GraphQLContext, GraphQLContextFactory, SchemaDirectiveTransformer } from '../classes';
 
 type SubschemaConfigOmittedProps = 'schema' | 'executor' | 'merge';
 type ServiceMixinSubschemaConfig = Omit<SubschemaConfig, SubschemaConfigOmittedProps>;
